refactor(panels): extract Miller columns rendering to src/view/panels.rs
- Move render_parent_pane(), render_current_pane(), and render_divider() to dedicated panels.rs module
- Create impl Heike block for panel rendering methods in view/panels.rs
- Simplifies app.rs by removing ~250 lines of rendering logic
- Maintains separation of concerns for UI layout
